backend/requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
numpy==2.0.1
pandas==2.2.2
py4j==0.10.9.7
pyarrow==17.0.0
PyJWT==2.9.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==71.1.0
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
